refactory deep copy and move it to utils
