+++ rust/agama-utils/Cargo.toml
@@ -8,6 +8,7 @@ edition.workspace = true
 agama-locale-data = { path = "../agama-locale-data" }
 agama-transfer = { path = "../agama-transfer" }
 async-trait = "0.1.89"
+camino = "1.2.1"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.140"
 serde_with = "3.14.0"
